update_manifest/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "update_manifest"
version = "0.0.0"
description = "Update Zephyr manifest with the latest module SHA"
authors = ["Sébastien Delpeuch <s.delpeuch@catie.fr>"]
packages = [
    { include = "update_manifest" }
]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pyyaml = "^6.0.1"
requests = "^2.31.0"

[tool.poetry.scripts]
update = "update_manifest.update:main"